simplify component
